+++ Runtime/IState.cs
@@ -83,6 +83,14 @@ namespace GameLovers.Statechart
 	}
 
 	/// <summary>
+	/// A transition state models a non-blocker situation in the execution of the <see cref="IStatechart"/> behavior.
+	/// Represents a non-blocking point on the state chart execution that automatically continues to the target state
+	/// </summary>
+	public interface ITransitionState : IStateEnter, IStateExit, IStateTransition
+	{
+	}
+
+	/// <summary>
 	/// A simple state models a situation in the execution of the <see cref="IStatechart"/> behavior.
 	/// Represents a blocking point on the state chart execution waiting for an event trigger to continue
 	/// </summary>
@@ -122,10 +130,10 @@ namespace GameLovers.Statechart
 	}
 
 	/// <summary>
-	/// A wait state is a blocking state that holds the <see cref="IStatechart"/> behavior.
+	/// A task wait state is a blocking state that holds the <see cref="IStatechart"/> behavior.
 	/// It waits for the completion of async defined <see cref="Task"/> to be completed to resume the state chart execution.
 	/// </summary>
-	public interface ITaskWaitState : IStateEnter, IStateExit
+	public interface ITaskWaitState : IStateEnter, IStateExit, IStateEvent
 	{
 		/// <summary>
 		/// Blocks the state behaviour until the given async <paramref name="taskAwaitAction"/> is completed.