+++ Runtime/Internal/TaskWaitState.cs
@@ -11,8 +11,9 @@ namespace GameLovers.Statechart.Internal
 	{
 		private ITransitionInternal _transition;
 		private Func<Task> _taskAwaitAction;
-		private bool _initialized;
+		private bool _triggered;
 		private bool _completed;
+		private uint _executionCount;
 
 		private readonly IList<Action> _onEnter = new List<Action>();
 		private readonly IList<Action> _onExit = new List<Action>();
@@ -20,16 +21,17 @@ namespace GameLovers.Statechart.Internal
 
 		public TaskWaitState(string name, IStateFactoryInternal factory) : base(name, factory)
 		{
-			_initialized = false;
+			_triggered = false;
 			_completed = false;
+			_executionCount = 0;
 		}
 
 		/// <inheritdoc />
 		public override void Enter()
 		{
-			_initialized = false;
+			_triggered = false;
 			_completed = false;
-
+			
 			for(int i = 0; i < _onEnter.Count; i++)
 			{
 				_onEnter[i]?.Invoke();
@@ -58,15 +60,6 @@ namespace GameLovers.Statechart.Internal
 			{
 				throw new InvalidOperationException($"The state {Name} is pointing to itself on transition");
 			}
-
-			foreach (var eventTransition in _events)
-			{
-				if (eventTransition.Value.TargetState?.Id == Id)
-				{
-					throw new InvalidOperationException(
-						$"The state {Name} with the event {eventTransition.Key.Name} is pointing to itself on transition");
-				}
-			}
 #endif
 		}
 
@@ -93,6 +86,21 @@ namespace GameLovers.Statechart.Internal
 		}
 
 		/// <inheritdoc />
+		public ITransition Event(IStatechartEvent statechartEvent)
+		{
+			if (statechartEvent == null)
+			{
+				throw new NullReferenceException($"The state {Name} cannot have a null event");
+			}
+
+			var transition = new Transition();
+
+			_events.Add(statechartEvent, transition);
+
+			return transition;
+		}
+
+		/// <inheritdoc />
 		public ITransition WaitingFor(Func<Task> taskAwaitAction)
 		{
 			_taskAwaitAction = taskAwaitAction ?? throw new NullReferenceException($"The state {Name} cannot have a null wait action");
@@ -109,10 +117,10 @@ namespace GameLovers.Statechart.Internal
 				return transition;
 			}
 
-			if (!_initialized)
+			if (!_triggered)
 			{
 				InnerTaskAwait(_taskAwaitAction);
-				_initialized = true;
+				_triggered = true;
 			}
 
 			return _completed ? _transition : null;
@@ -120,11 +128,19 @@ namespace GameLovers.Statechart.Internal
 
 		private async void InnerTaskAwait(Func<Task> taskAwaitAction)
 		{
+			var currentExecution = _executionCount;
+
+			_executionCount++;
+			
 			await taskAwaitAction();
 
 			_completed = true;
-			
-			_stateFactory.Data.StateChartMoveNextCall(null);
+
+			// Checks if the state didn't exited from an outsource trigger (Nested State) before the Task was completed
+			if (_executionCount == currentExecution)
+			{
+				_stateFactory.Data.StateChartMoveNextCall(null);
+			}
 		}
 	}
 }