+++ Runtime/Internal/WaitActivity.cs
@@ -9,6 +9,7 @@ namespace GameLovers.Statechart.Internal
 	internal interface IWaitActivityInternal : IWaitActivity
 	{
 		uint Id { get; }
+		uint ExecutionCount { get; set; }
 		bool IsCompleted { get; }
 		void Reset();
 	}
@@ -18,13 +19,14 @@ namespace GameLovers.Statechart.Internal
 	{
 		private static uint _idRef;
 
-		private Action<IStatechartEvent> _onComplete;
-		private Action<uint> _innerOnComplete;
-
+		private readonly Action<IStatechartEvent> _onComplete;
+		private readonly Action<uint> _innerOnComplete;
 		private readonly Dictionary<uint, bool> _activities = new Dictionary<uint, bool>();
 
 		/// <inheritdoc />
 		public uint Id { get; private set; }
+		/// <inheritdoc />
+		public uint ExecutionCount { get; set; }
 		/// <inheritdoc />
 		public bool IsCompleted { get; private set; }