+++ Tests/Editor/StatechartTaskWaitTest.cs
@@ -24,6 +24,7 @@ namespace GameLoversEditor.StateChart.Tests
 		}
 		
 		private readonly IStatechartEvent _event1 = new StatechartEvent("Event1");
+		private readonly IStatechartEvent _event2 = new StatechartEvent("Event2");
 
 		private IMockCaller _caller;
 		private bool _blocker;
@@ -36,7 +37,7 @@ namespace GameLoversEditor.StateChart.Tests
 		}
 
 		[Test]
-		public async void BasicSetup()
+		public void BasicSetup()
 		{
 			var statechart = new Statechart(factory => SetupTaskWaitState(factory, TaskWaitAction));
 
@@ -46,18 +47,127 @@ namespace GameLoversEditor.StateChart.Tests
 			_caller.DidNotReceive().OnTransitionCall(1);
 			_caller.DidNotReceive().OnTransitionCall(2);
 			_caller.Received().InitialOnExitCall(0);
-			_caller.Received().StateOnEnterCall(1);
-			_caller.DidNotReceive().StateOnExitCall(1);
+			_caller.Received().StateOnEnterCall(0);
+			_caller.DidNotReceive().StateOnExitCall(0);
 			_caller.DidNotReceive().FinalOnEnterCall(0);
 
 			_blocker = false;
 
-			await Task.Delay(11);
+			Task.Run(async () =>
+			{
+				await Task.Delay(10);
+				
+				_caller.Received().OnTransitionCall(1);
+				_caller.DidNotReceive().OnTransitionCall(2);
+				_caller.Received().StateOnExitCall(0);
+				_caller.Received().FinalOnEnterCall(0);
+			});
+		}
 
-			_caller.Received().OnTransitionCall(1);
+		[Test]
+		public void EventTrigger()
+		{
+			var statechart = new Statechart(factory => SetupTaskWaitState(factory, TaskWaitAction));
+
+			statechart.Run();
+
+			_caller.Received().OnTransitionCall(0);
+			_caller.DidNotReceive().OnTransitionCall(1);
 			_caller.DidNotReceive().OnTransitionCall(2);
+			_caller.Received().InitialOnExitCall(0);
+			_caller.Received().StateOnEnterCall(0);
+			_caller.DidNotReceive().StateOnExitCall(0);
+			_caller.DidNotReceive().FinalOnEnterCall(0);
+
+			statechart.Trigger(_event1);
+			
+			_blocker = false;
+
+			Task.Run(async () =>
+			{
+				await Task.Delay(10);
+				
+				_caller.Received().OnTransitionCall(1);
+				_caller.Received().OnTransitionCall(2);
+				_caller.Received().StateOnExitCall(0);
+				_caller.Received().FinalOnEnterCall(0);
+			});
+		}
+
+		[Test]
+		public void NestState_OuterEventTriggerAndTaskComplete_NoTransition()
+		{
+			var statechart = new Statechart(InternalSetupNest);
+
+			statechart.Run();
+
+			_caller.Received().OnTransitionCall(0);
+			_caller.DidNotReceive().OnTransitionCall(1);
+			_caller.DidNotReceive().OnTransitionCall(2);
+			_caller.Received().OnTransitionCall(3);
+			_caller.DidNotReceive().OnTransitionCall(4);
+			_caller.DidNotReceive().OnTransitionCall(5);
+			_caller.Received().InitialOnExitCall(0);
+			_caller.Received().InitialOnExitCall(1);
+			_caller.Received().StateOnEnterCall(0);
+			_caller.Received().StateOnEnterCall(1);
+			_caller.DidNotReceive().StateOnExitCall(0);
+			_caller.DidNotReceive().StateOnExitCall(1);
+			_caller.DidNotReceive().FinalOnEnterCall(0);
+			_caller.DidNotReceive().FinalOnEnterCall(1);
+
+			statechart.Trigger(_event2);
+
+			_caller.DidNotReceive().OnTransitionCall(1);
+			_caller.DidNotReceive().OnTransitionCall(2);
+			_caller.DidNotReceive().OnTransitionCall(4);
+			_caller.Received().OnTransitionCall(5);
+			_caller.Received().StateOnExitCall(0);
 			_caller.Received().StateOnExitCall(1);
-			_caller.Received().FinalOnEnterCall(0);
+			_caller.DidNotReceive().FinalOnEnterCall(0);
+			_caller.Received().FinalOnEnterCall(1);
+			
+			_blocker = false;
+
+			Task.Run(async () =>
+			{
+				await Task.Delay(10);
+				
+				_caller.DidNotReceive().OnTransitionCall(1);
+				_caller.DidNotReceive().OnTransitionCall(2);
+				_caller.DidNotReceive().OnTransitionCall(4);
+				_caller.Received(1).OnTransitionCall(5);
+				_caller.Received(1).StateOnExitCall(0);
+				_caller.Received(1).StateOnExitCall(1);
+				_caller.DidNotReceive().FinalOnEnterCall(0);
+				_caller.Received(1).FinalOnEnterCall(1);
+			});
+
+			void InternalSetupNest(IStateFactory factory)
+			{
+				SetupNest(factory, _event2, InnerSetupTaskWaitState, true, false);
+			}
+
+			void InnerSetupTaskWaitState(IStateFactory factory)
+			{
+				SetupTaskWaitState(factory, TaskWaitAction);
+			}
+		}
+
+		[Test]
+		public void MissingTaskWaiter_ThrowsException()
+		{
+			Assert.Throws<MissingMethodException>(() => new Statechart(factory =>
+			{
+				var initial = factory.Initial("Initial");
+				var waiting = factory.TaskWait("Waiting");
+
+				initial.Transition().OnTransition(() => _caller.OnTransitionCall(0)).Target(waiting);
+				initial.OnExit(() => _caller.InitialOnExitCall(0));
+
+				waiting.OnEnter(() => _caller.StateOnEnterCall(1));
+				waiting.OnExit(() => _caller.StateOnExitCall(1));
+			}));
 		}
 
 		[Test]
@@ -81,7 +191,7 @@ namespace GameLoversEditor.StateChart.Tests
 		{
 			while (_blocker)
 			{
-				await Task.Delay(10);
+				await Task.Delay(1);
 			}
 		}
 
@@ -94,11 +204,30 @@ namespace GameLoversEditor.StateChart.Tests
 			initial.Transition().OnTransition(() => _caller.OnTransitionCall(0)).Target(waiting);
 			initial.OnExit(() => _caller.InitialOnExitCall(0));
 
-			waiting.OnEnter(() => _caller.StateOnEnterCall(1));
+			waiting.OnEnter(() => _caller.StateOnEnterCall(0));
 			waiting.WaitingFor(waitAction).OnTransition(() => _caller.OnTransitionCall(1)).Target(final);
-			waiting.OnExit(() => _caller.StateOnExitCall(1));
+			waiting.Event(_event1).OnTransition(() => _caller.OnTransitionCall(2));
+			waiting.OnExit(() => _caller.StateOnExitCall(0));
 
 			final.OnEnter(() => _caller.FinalOnEnterCall(0));
+		}
+
+		private void SetupNest(IStateFactory factory, IStatechartEvent eventTrigger, Action<IStateFactory> nestSetup,
+		                       bool executeExit, bool executeFinal)
+		{
+			var initial = factory.Initial("Initial");
+			var nest = factory.Nest("Nest");
+			var final = factory.Final("final");
+
+			initial.Transition().OnTransition(() => _caller.OnTransitionCall(3)).Target(nest);
+			initial.OnExit(() => _caller.InitialOnExitCall(1));
+
+			nest.OnEnter(() => _caller.StateOnEnterCall(1));
+			nest.Nest(nestSetup, executeExit, executeFinal).OnTransition(() => _caller.OnTransitionCall(4)).Target(final);
+			nest.Event(eventTrigger).OnTransition(() => _caller.OnTransitionCall(5)).Target(final);
+			nest.OnExit(() => _caller.StateOnExitCall(1));
+
+			final.OnEnter(() => _caller.FinalOnEnterCall(1));
 		}
 	}
 }